requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Pillow==10.1.0
imagehash==4.3.1
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
APScheduler==3.10.4
PyPDF2==3.0.1
reportlab==4.0.7
jellyfish==0.9.0